+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Es ersetzt die **Release Notes** aus den Anfangstagen von jFachwert und orientie
 genauso wie an [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Patch-Versionen nicht extra aufgeführt, sondern in der Minor-Version.
 
+## [Unreleased]
+- Encoding-Probleme unter Java 17 behoben
+
 ## [4.3.0] - 2022-08-27
 ### Added
 - Text.of(..) akzeptiert jetzt als 2. Parameter einen Zeichensatz- Text-Konvertierung nach ASCII (und andere Zeichensätze) wird unterstützt